+++ src/components/Trends/TrendsTable/TrendsTable.module.scss
