+++ src/components/Trends/TrendsTable/TrendsTables.module.scss
